# server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Pocket Tarot API
  version: 0.1.0
  description: Pocket Tarot Express API for auth, readings, profile, admin maintenance, and API documentation.
servers:
  - url: /api/v1
    description: API base path
...
paths:
...
  /api/v1/admin/users:
    get:
      tags: [Admin]
      summary: List users for internal maintenance.
      security:
        - adminKey: []
      parameters:
        - in: query
          name: deleted
          schema:
            type: string
            enum: [false, true, all]
        - in: query
          name: isActive
          schema:
            type: boolean
        - in: query
          name: email
          schema:
            type: string
      responses:
        "200":
          description: User list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/admin/users/{id}:
    get:
      tags: [Admin]
      summary: Get one user for internal maintenance.
      security:
        - adminKey: []
      parameters:
        - $ref: "#/components/parameters/ObjectId"
      responses:
        "200":
          description: User.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/v1/admin/users/{id}/activation:
    patch:
      tags: [Admin]
      summary: Activate or deactivate a user.
      security:
        - adminKey: []
      parameters:
        - $ref: "#/components/parameters/ObjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActivationRequest"
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/v1/admin/users/{id}/deletion:
    patch:
      tags: [Admin]
      summary: Soft delete or restore a user.
      security:
        - adminKey: []
      parameters:
        - $ref: "#/components/parameters/ObjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeletionRequest"
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
components:
  securitySchemes:
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  parameters:
    ObjectId:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: MongoDB ObjectId.
  responses:
    ErrorResponse:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              enum:
                - UNAUTHENTICATED
                - EMAIL_NOT_VERIFIED
                - PROFILE_NOT_FOUND
                - ACCOUNT_NOT_ACTIVE
                - ACCOUNT_DELETED
                - VALIDATION_ERROR
                - DAILY_READING_NOT_FOUND
                - LLM_UNAVAILABLE
                - LLM_TIMEOUT
                - LLM_INVALID_RESPONSE
                - ADMIN_UNAUTHORIZED
                - EMAIL_ALREADY_REGISTERED
                - NOT_FOUND
                - INTERNAL_ERROR
            message:
              type: string
            details:
              type: object
...
    User:
      type: object
      required: [id, email, displayName, providerIds, isActive, createdAt]
      properties:
        id:
          type: string
        firebaseUid:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        providerIds:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        activatedAt:
          type:
            - string
            - "null"
          format: date-time
        lastLoginAt:
          type:
            - string
            - "null"
          format: date-time
        deletedAt:
          type:
            - string
            - "null"
          format: date-time
    UserResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [user]
          properties:
            user:
              $ref: "#/components/schemas/User"
...
    UserListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
...
    UpdateActivationRequest:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean
    UpdateDeletionRequest:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean